Started exchanging info between process
Working for up/down/left/right, but not diagonal.

Have not update edges of segment that each process is responsible for.
Will try to skip the copying and write directly to the grid matrix.
Will also try to wrap the matrix in an additional layer to make it easier to compute.